requirements.txt
pandas
keras
nltk
flask
tensorflow
keyBERT
sentence_transformers
transformers==4.18.0
torch==1.11.0
sentencepiece
langdetect
newspaper3k
seaborn
sklearn
scipy
matplotlib
